requirements.txt
numpy==1.26.3
pillow==10.2.0
pytesseract==0.3.13
pyspellchecker==0.8.2